moved FakeDataSource.kt into network package
