+++ app/src/main/java/com/digitalarchitects/rmc_app/network/FakeDataSource.kt
@@ -1,4 +1,4 @@
-package com.digitalarchitects.rmc_app.data
+package com.digitalarchitects.rmc_app.network
 
 import com.digitalarchitects.rmc_app.model.User
 import com.digitalarchitects.rmc_app.model.Vehicle
