+++ app/src/test/java/com/digitalarchitects/rmc_app/FakeRmcApiService.kt
@@ -1,6 +1,6 @@
 package com.digitalarchitects.rmc_app
 
-import com.digitalarchitects.rmc_app.data.FakeDataSource
+import com.digitalarchitects.rmc_app.network.FakeDataSource
 import com.digitalarchitects.rmc_app.model.User
 import com.digitalarchitects.rmc_app.model.Vehicle
 import com.digitalarchitects.rmc_app.network.RmcApiService
